requirements.txt
boto3>=1.28
botocore>=1.31
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
geopandas>=0.14
shapely>=2.0
scikit-learn>=1.3
nltk>=3.8
faiss-cpu>=1.7
fastapi>=0.104
uvicorn>=0.24
jinja2>=3.1
